Build explorer stack images in CI and build them on GCloud (#4888)
## Motivation

If we want to get the explorer stack working on the testnet, we need to
also build their images in CI

## Proposal

Build them in CI, and also build them in GCloud instead of Github
runners, as it's supposed to be faster and maybe even cheaper.

## Test Plan

Added a test commit to build test images in the PR's CI to make sure
things work, then will remove it.