Crash exporter if block processor task crashes (#4900)
## Motivation

Right now, if the block processor returns an error, since we await the
service first, we'll never see that error unless the service receives
some shutdown signal or itself finishes for some reason, which in most
cases doesn't happen.
If the block processor returns an error, we should see that error and
propagate it, even if the service is still running and receiving
notifications.

## Proposal

poll both futures at once with `select!` so that if the block processor
returns an error, we can panic as we should

## Test Plan

TBD