Web client: make block_hash optional arg of query fn (#4904)
Backport of #4905 

## Motivation

In previous PRs we introduced an option to query app's state at a
specific block (hash). It made the `query` method NOT backwards
compatible b/c it required the new argument to be present.

## Proposal

Make argument optional. This allows `wasm-bindgen` to generate a code
that does not fail when argument is not passed in.

## Test Plan

Manual. I tested the metamask demo – which doesn't supply the argument –
and it worked.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.


## Links


- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)