Add gRPC keepalive configuration to detect stale connections (#4908)
## Motivation

Seems that `connect_lazy` doesn't reconnect automatically unless we
specify a keep alive https://github.com/hyperium/tonic/issues/1254
There's some evidence we might be seeing that on the testnet on the
proxy, as the user from the issue is seeing the same errors we're
seeing.

## Proposal

Add keep alives so that they can expire and `connect_lazy` can reconnect
automatically.

## Test Plan

We can reproduce this on the testnet. We can deploy to `conway-test`
just to make sure it doesn't break anything (run e2e tests, for
example), and then deploy to `testnet-conway`.