Add the `root_key` to the `MissingEntries` error. (#4916)
## Motivation

We have some error of `MissingEntries` occurring in the logs. We do not
know what is the relevant root key.
So, we add it here.

## Proposal

The following has been done:
* The `fn root_key(&self) -> Result<Vec<u8>, Self::Error>` is added to
the trait definition. We need a `Result` because of the deserialization
of the keys in RocksDB and StorageService.
* For RocksDB, ScyllaDB, DynamoDB, and StorageService, this did not add
any new data.
* The key is not added to the error since this would force some clone
operation.



## Test Plan

The CI.

## Release Plan

The change could be backported to TestNet Conway.

## Links

None.