Add sqlite3 to Indexer image (#4919)
## Motivation

This is useful for debugging, in case we want to go into the pod and see
what's going on with the sqlite db.
We can't use the debug docker image for this, because sharing the same
disk as the main container is trickier to setup.

## Proposal

Add `sqlite3` to the image.

## Test Plan

CI, this should be pretty harmless, and I've checked that `sqlite3` is
possible to be installed there.

## Release Plan

- These changes have already been backported to the latest `testnet`
branch