Add gRPC keepalive configuration to detect stale connections (#4918)
## Motivation

This is a forward port of #4908.

Seems that `connect_lazy` doesn't reconnect automatically unless we
specify a keep alive https://github.com/hyperium/tonic/issues/1254
There's some evidence we might be seeing that on the testnet on the
proxy, as the user from the issue is seeing the same errors we're
seeing.

## Proposal

Add keep alives so that they can expire and `connect_lazy` can reconnect
automatically.

## Test Plan

This has been deployed to `testnet-conway`

## Release Plan

- These changes have already been backported to the latest `testnet`
branch