Implement `Allocative` for the protocol views. (#4687)
## Motivation

Tracking the size of the objects being used is important for considering
the memory usage and resolving the OOM.

## Proposal

The `Allocative` trait is implemented for the views, and the code is
followed from there.

Several design decisions were made:
* For the Weighted distribution, the entries are dropped.
* In the views, the context is not taken into account of memory usage.
* For the keys of Ed25519, Secp256k1, EvmSecp256k1, the sizes are
written explicitly.

## Test Plan

The CI.

Adding in `process_confirmed_block` the following 
```rust
tracing::info!("process_confirmed_block, size_of_unique={}", allocative::size_of_unique(&self.chain));
```

works as expected. How to implement more systematically is open to
discussion.

## Release Plan

The goal is to add it to the TestNet Conway.

## Links

None.