Don't pass empty string, it's optional (#4929)
## Motivation

We are seeing errors in the console, coming from the Linera web client,
about CryptoHash expecting 32-bytes but getting 0.

## Proposal

This happens b/c frontends pass empty string (`''`) to a method
(`query`) that expects `Option<CryptoHash>` and assuming that if it's
`Some(_)` it must be correct slice.

After fixing `query` in #4904 I forgot to update this.

## Test Plan

Manual. Verified that with this fix we no longer get the errors.

## Release Plan

- Release web demos to the frontend
- Backport fix to `main`

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)