Build explorer stack images in CI and build them on GCloud (#4933)
## Motivation

This is a forward port of #4888.

If we want to get the explorer stack working on the testnet, we need to
also build their images in CI

## Proposal

Build them in CI, and also build them in GCloud instead of Github
runners, as it's supposed to be faster and maybe even cheaper.

## Test Plan

This was tested on the `testnet_conway` branch.