When first request fails, start subsequent ones in parallel with increasing delay. (#4913)
## Motivation

We observed that when the first request failed, the failure was
broadcasted to all waiting peers. This slowed down the process and
didn't use alternative peers we expect to have the data.

## Proposal

Detect if request fails and if so, try all alternative peers before
erroring. We spawn a retry operation for every alternative peer with
ever-increasing delay, delayed by `75ms` by default.

## Test Plan

CI (a test was added for this case).

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK,

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)