Implement a function `has_empty_storage` in `BaseRuntime`. (#4530)
## Motivation

For the purpose of testing if a contract exists in EVM, we introduce the
functionality of testing if a contract
has empty storage.

## Proposal

The following is done:
* The functionality is added to the `BaseRuntime`.
* It uses the `TotalStorageSize` in the `execution_state_actor.rs`.
* It uses the `view.total_size()` of the `KeyValueStoreView`.
* The functionality is provided to the Wasm smart contracts.

Several points have to be noted:
* The Wasm smart contracts are persistent. Only at the end with the
`self.state.save().await` does their state become observable. This point
is exemplified in the example test.
* So, if a contract is writing its state only at the `fn finalize` step,
then we cannot see that its state is non-trivial only at the next block.
* The designed use case for this feature is EVM smart contracts.
Fortunately, for those contracts, we are writing the state directly to
the storage and we are not waiting for `fn finalize`. So, that is fine
here.

## Test Plan

The CI.

For testing the `has_empty_storage`, we implemented separately some
end-to-end tests.
They have not been put in the PR, since you need two contracts:
* one that gets instantiated
* and one that does the `has_empty_storage`.

## Release Plan

This does not appear to be TestNet breaking.

## Links

The test is implemented in
[test_has_empty_state](https://github.com/MathieuDutSik/linera_end_to_end_tests/tree/main/test_has_empty_state)