Allow zero key length in key and prefix for `DynamoDb`. (#4893)
## Motivation

Recently, the operation `find_keys_by_prefix(&[])` has emerged as an
operation on stores. This operation
does not work on `DynamoDb` since the prefix has to be non-empty as well
as the key.

## Proposal

Add a `[1]` as prefix.

## Test Plan

The CI should catch the bugs. The `run_reads` has been updated for that.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
It is not that important since DynamoDB is not used on TestNet Conway.

## Links

None.