Restructure the CI to diminish load. (#4935)
## Motivation

We have high CI expenses that should be diminished as much as possible.

## Proposal

The following steps are done:
* Move the Ethereum tests to a separate file that is triggered only on
`linera-execution`, `linera-ethereum` and other relevant directory. Just
like for DynamoDB, ScyllaDB.
* Harmonize the dynamodb.yml, scylladb.yml
* Remove a `DOCKER_COMPOSE_WAIT` that is wrongly put.
* Put the `remote-net-test`, `remote-kubernetes-net-test`, and
`benchmark-test` as merge group just like docker-compose and other 3
tests.
* Add the `concurrency` group to the `long_faucet_chain_test.yml`.

## Test Plan

The CI.

## Release Plan

Some of the changes can be backported to the TestNet Conway.

## Links

None.