First request should be subject to a delayed alternative. (#4951)
## Motivation

`RequestsScheduler` sends a request to a first peer and, if it fails or
time out, enter `try_staggerred_parallel` function which will try new
requests after ever-increasing delay (but before the requests time out
or fail). This is inconsistent and vulnerable when first request is sent
to a slow/unavailable/malicious peer.

## Proposal

Make the logic consistent: first request is subject to the same
retry-after-delay logic as the rest. If delay fires first, a new request
(to an alternative peer) is started and delay updated. If any request
fails w/ an error – a new request is started.

## Test Plan

A unit test already exists for this.

## Release Plan

- These changes should be deployed to testnet and then backported to
`main`.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)