Synchronize balances between Linera and the EVM. (#4274)
## Motivation

Linera has its native-token (do we call it lineras?), and we need to
have synchronicity between it and the balance
of the EVM. This is harder to achieve

Fixes #3756 

## Proposal

In the EVM world, there is complete identity between transferring ethers
and calling a function. This is not so in Linera.

Therefore, some changes have to be made for it to work:
* Linera amounts are on 128 bits while EVM ethers are on 256 bits. But
that is not a problem since, in effect, only 85 bits are ever used. But
this introduces a conversion function.
* Now the `transact_commit` takes an amount being transmitted and the
function call. Just like in the EVM world.
* We introduce a `get_mutation` function that takes an amount, an
operation, and encapsulates it into a single `Vec<u8>`.
* This is followed when operations are transmitted to another function.
* Currently, the `fn transfer` function works in the sense that the
beneficiary receives the ethers on the next block. Since the
beneficiaries of transfers are on the same chain as the source, we can
make it sync.
* When instantiating a new contract, we can start with some value. But
this is only for the creator chain. For other chains, nothing is
received.

Other changes being made:
* Better tracking of occurring errors.
* A type `EvmInstantiation` is introduced that contains the native token
at instantiation and the command for instantiating.
* Move some code into `inputs.rs`.

Further needed work:
* The beneficiary of a transfer operation can be an Externally Owned
Account or a Smart Contract. The problem is in distinguishing between
the two. In this PR we use it by the size of the operation; if null,
then it is an EOA. But this is not strictly exact. What matters in the
end is whether the corresponding blob exists, since if we call an
application that does not yet exist but has a blob, then it is
instantiated and executed.

## Test Plan

The CI.
The test that demonstrates the functionality is
`test_evm_end_to_end_balance_and_transfer`.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.