Replace `wasm_thread` with `web-thread` (#4959)
## Motivation

Our hacked-up version of `wasm_thread` isn't a great fit for our use
cases.

Firstly, `wasm_thread` aims to mimic the API of `std::thread`. This was
never sufficient for us as we need to send non-`Send` messages across
threads actively using `postMessage`, which `std::thread` has no
equivalent for. We patched `wasm_thread` but it was at odds with the
point of the package.

Then, in order to use `wasm_thread` without threading a shim script URL
through the code down into `linera-base`, we dynamically got the URL of
the `wasm-bindgen` shim script using `import.meta.url`. This is poorly
(and inconsistently) supported by bundlers, which want to know what the
URL will be used for; but since the `Worker` construction in
`wasm_thread` is done in Rust code they cannot analyze it. This resulted
in our `@linera/client` library being rather fragile under different
bundlers, with workarounds required to ensure they don't interfere with
the paths of the various files.

## Proposal

Introduce the
[`web-thread`](https://docs.rs/web-thread/latest/web_thread/) library, a
simplified equivalent to `wasm_thread` that has a Web-first API,
including a trait for types that can be passed by `postMessage`
(including [transferable
objects](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Transferable_objects)),
and whose worker logic is primarily implemented in JavaScript, making it
much more comprehensible to bundlers.

This allows us:
- to communicate both `Send` and `Post` objects with long-running
threads, as is required by
https://github.com/linera-io/linera-protocol/issues/3651 and eventually
https://github.com/linera-io/linera-protocol/issues/2927
- to remove the bundling caveats from Linera documentation
- to support WebPack
- to delete `linera_base::task` and the conditional compilation therein

## Test Plan

The `web-thread` library itself has been tested with Vite as well as
WebPack (via Next.js with the `--webpack` flag), and includes example
applications to that effect.

I have manually tested our two Web applications in this repository.

I defer to CI for full tests of the native pathways.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Fixes https://github.com/linera-io/linera-protocol/issues/2809.
- First step towards
https://github.com/linera-io/linera-protocol/issues/3651.
- [`web-thread` docs](https://docs.rs/web-thread/latest/web_thread/)

### Turbopack

Turbopack is currently not supported due to an open issue when
processing cyclic dependencies. See the following discussions for more
information:

* https://github.com/vercel/next.js/issues/85119
* https://github.com/vercel/next.js/discussions/77102
* https://github.com/emscripten-core/emscripten/issues/20580