Return contexts with correct base keys from views (#4983)
## Motivation

The derived implementation of `View::context` currently returns just the
context of the first sub-view of the view. This means that if a view A
has a sub-view B, and view B has a sub-view C, when A calls `context()`
on B and assumes that its own base key is eg. a byte shorter than what
it gets as a result, it will in fact get a context with C's base key,
which may lead to wrong results if B also appends some bytes to the base
key before passing the context on to C (which happens in hashable
wrappers, for example!).

## Proposal

Make the implementation of `context()` in all views correctly trim the
base key contained in the context before returning the value.

This unfortunately introduces the necessity of cloning the context and
returning it by value instead of by reference.

An alternative (but somewhat more complicated) solution would be to
require _all_ views to store their inner contexts, and make the `View`
derive macro require a `context` field in all structs for which `View`
is derived.

## Test Plan

CI will catch regressions.

This was also manually confirmed to fix a bug that surfaced in #4977.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- This bug may be causing some issues on the testnet, but the fix cannot
be safely backported without a complex migration.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)