[testnet] Allow filtering incoming messages by application. (#4989) (#5003)
Backport of #4989.

## Motivation

It's useful to filter out spam messages on a node service that e.g. is
running a chain specifically for one application.

## Proposal

Allow filtering in two ways:
* All bundles that contain _at least one_ message for a whitelisted app.
* All bundles that contain _only_ messages for whitelisted apps.

The latter is stricter, but the former can be useful e.g. for apps that
create other apps and then call them.

## Test Plan

A test for the new filtering options was added. (Claude)

## Release Plan

- These changes should be backported to `testnet_conway`, then
    - be released in a new SDK.

## Links

- PR to main: #4989.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)