Improve the speed of `pre_save` in `BucketQueueView`. (#5011)
## Motivation

In the PR https://github.com/linera-io/linera-protocol/pull/4953 and
https://github.com/linera-io/linera-protocol/pull/4961
we restructured the API. Great care was made to be correct, but not so
much with speed.

## Proposal

Avoid both clones in `pre_save`: one for `stored_buckets`, one for
`new_backValues`.

## Test Plan

The CI.
The benchmarks do show improvements.

## Release Plan

Can be backported to TestNet Conway.

## Links

None.