Restructure the `MultiPartitionBatch` and the keys. (#5019)
## Motivation

Due to the effort to make the database migration as good as possible,
the database schema of `testnet_conway` ends up being better than the
one of `main`.

## Proposal

Backport the changes from the `testnet_conway` to `main` (reverse of the
usual way).
This covers two changes:
* The `MultiPartitionBatch` can support several `key/values` on the sme
`root_key`.
* The keys are now no longer `[0]` and `[1]` which is a little better.

## Test Plan

The CI.

## Release Plan

Unify the `testnet_conway` and `main` is the end result of the PR.

## Links

None.