Cleanups and minor fixes to rounds, ownership and leader calculation. (#5033)
## Motivation

Much of the code related to rounds and chain owners is hard to read and
sometimes inconsistent.

* The logic for computing the round leader is currently repeated in
three places.
* The client would fail to recognize that fallback owners are owners
unless they are the current leader.
* The client could fail in some edge cases because
`round_for_new_proposal` is used to determine the validation round
oracle value.

## Proposal

* Deduplicate round leader calculation.
* Extract some code into `ChainOwnership` methods; rename `verify_owner`
to `is_owner`.
* Super owners _can_ propose in all except validator rounds, because in
fallback mode the validators are supposed to override them and receive
mandatory messages. They _should_ avoid proposing in other owners'
single-leader rounds, however, so that will be the client's default
behavior.
* Fix the fallback ownership check in the client.
* Fix the validation round oracle value computation: Even if an error is
returned because the current round is `Fast` and we should not propose
in any later round, set the value to `None` and continue.

## Test Plan

A regression test for the validation round oracle issue was added.

## Release Plan

- These changes could be backported to `testnet_conway`, then
    - be released in a new SDK,
    - be released in a validator hotfix.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)