Modify tests for historical hashing (#5040)
## Motivation

After the `ExecutionStateView` switches to historical hashing, it will
be very hard to manually compute a hash of a given execution state,
since it will depend not on the state itself, but on its history. That
will break a lot of the chain worker tests.

## Proposal

Make the chain worker tests create certificates (and thus compute hashes
inside) by executing the corresponding blocks on an auxiliary worker,
then assert that the execution outcomes are as expected.

Two tests are being removed:
- `test_handle_certificate_with_open_chain`, as it has only been testing
sending a transfer to a child chain,
- `test_handle_certificate_wrong_owner`, as it tested a very artificial
scenario of a wrong owner being used to initialize a state used to
generate a certificate, and the state hash was the only part of the
certificate being affected, which we can no longer test for.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- prerequisite for #4977 
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)