Ignore missing cert errors if it's expected. Load it from env (#5039)
## Motivation

We know that on testnet some certs are missing so it doesn't make sense
to raise critical errors on them which crash the exporter.

## Proposal

Load, from an env var `LINERA_KNOWN_MISSING_CERTS`, list of hashes that
are expected to be missing. If we fail to read a certificate that is on
that list, we don't treat it as an error.

## Test Plan

<!--
Explain how you made sure that the changes are correct and that they
perform as intended.

Please describe testing protocols (CI, manual tests, benchmarks, etc) in
a way that others
can reproduce the results.
-->

## Release Plan

- Testnet only.
-
## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)