Remove the memory limit from Ethereum. (#5061)
## Motivation

Ethereum has a strict limitation to 24576 bytes which forces the use of
multiple contracts for effective running of Ethereum functionalities.

## Proposal

The limit is set to the `maximum_blob_size`.

This requires adding the functionality to the `BaseRuntime`. We only
added the needed function and did not add the corresponding Wasm access.
We also did not add full access to `ResourceControlPolicy`. This would
expose the internals and lead to an abstraction leakage.

For testnet, this sets the limit to 1M for the contract, which is a
radical change coming from Ethereum.

## Test Plan

The CI is the normal way.
Locally, we had proof that these changes do address the limit.

## Release Plan

Whenever Ethereum is deemed acceptable for TestNet, we should broadcast
this change.

## Links

None.