Measure operation/message execution latency in microseconds (#5007)
## Motivation

Our latency for operation/message execution is actually in the
microseconds in the benchmarks. So it makes sense to actually track this
in microseconds instead, to avoid dealing with super small numbers

## Proposal

Added support for measuring latency in different units, and started
measuring operation/message execution in microseconds.

## Test Plan

Deployed a network with this, saw the updated metric

## Release Plan

- Nothing to do / These changes follow the usual release cycle.