Node service plugins (#5072)
On a private repository Mathieu started working on an alternative node
service with a plugin system for external processes that want to
interact with apps and commit operations.

Add the plugin system to the node service itself. (The task processor
part done by Claude, but reviewed.)

This PR is an extension of #5056

CI
(Claude added a test; to extend that we'd have to write an example app
that requests actions from a plugin.)

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK,

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Co-authored-by: Andreas Fackler <andreas.fackler@linera.io>