Use `QueueView` instead of `BucketQueueView`. (#5096)
## Motivation

In the benchmarks we are seeing that a lot of time is spent in
`BucketQueueView::pre_save`. My theory is that it's serializing the list
of buckets, which takes O(N) time depending on the number N _of total
entries in the queue_! (_Not_ only on _newly added_ entries!)

```
  | Queue Type      | Time for 100 incremental saves (10,000 existing elements) |
  |-----------------|-----------------------------------------------------------|
  | QueueView       | 33.6 - 34.3 µs                                            |
  | BucketQueueView | 171.5 - 179.3 µs                                          |
```

(And that's with optimally-filled 100-item buckets for the first 10000
items!)

## Proposal

Use `QueueView` instead, so it scales better. We can switch back once we
fix the scalability issues of `BucketQueueView`.

## Test Plan

CI

A benchmark was added. We should also try this out with the large-scale
benchmark in a real network.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)