Retry and warn if validators return `InvalidTimestamp`. (#5090)
## Motivation

Validators return `InvalidTimestamp` errors if a proposed block's
timestamp is too far in the future. (For a short grace period, 500 ms by
default, they wait.)

Some users are getting this error, possibly because their local clock is
not in sync with the validators'.

## Proposal

Whenever we get `InvalidTimestamp` from a validator, we wait and retry.

If the local clock is the problem and not the validators', this can
cause long delays, of course. So we print a warning log message as soon
as a validity threshold has returned this error.

## Test Plan

A test was added.

## Release Plan

- We should think about whether and how some of this can be backported
and released in a new SDK. But it's not trivial, since this PR changes
the `NodeError` type.

## Links

- Related to https://github.com/linera-io/linera-protocol/issues/5023.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)