Follow-only chains (#5101)
## Motivation

All chains in the wallet are currently fully synchronized, including
their sender and publisher chains. But in use cases where the user only
wants to _read_ from that chain and not _write_ to it (i.e. never make a
block proposal, e.g. because they don't even own it), sender and
publisher chains are not needed, and neither are the chain manager
values.

## Proposal

Add a new listening mode and flag that marks chains as "follow-only".
For those chains, only their own confirmed blocks are synchronized.

The `follow-chain` and `forget-keys` commands automatically set that
flag, and `assign` unsets it.

## Test Plan

A test was added.

## Release Plan

- These changes could possibly be backported to `testnet_conway`, then
    - be released in a new SDK.

## Links

- Closes #5098.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)