Port the `ListeningMode` enum to the testnet branch (#5111)
## Motivation

As a part of the sparse event chains PR (#4466), an enum controlling how
the listener listens to chains was introduced. This mechanism will now
be useful for some features being ported to the testnet branch.

## Proposal

Port the enum and its usage to the testnet branch. Omit the `EventsOnly`
variant for now. This makes the changes in this PR kind of pointless by
themselves, but they will make it easier to introduce further changes in
the future.

## Test Plan

CI

## Release Plan

- These changes should be released in a new SDK

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)