[testnet] Retry and warn if validators return InvalidTimestamp. (#5090) (#5108)
Backport of #5090.

## Motivation

Validators return `InvalidTimestamp` errors if a proposed block's
timestamp is too far in the future. (For a short grace period, 500 ms by
default, they wait.)

Some users are getting this error, possibly because their local clock is
not in sync with the validators'.

## Proposal

Whenever we get `InvalidTimestamp` from a validator, we wait and retry.

If the local clock is the problem and not the validators', this can
cause long delays, of course. So we print a warning log message as soon
as a validity threshold has returned this error.

To backport this to `testnet_conway` in a backwards-compatible way,
instead of adding a variant to `NodeError` (which would cause
deserialization errors in clients that weren't updated), I changed the
error string and parse it. That way, even if the validator is updated
but the client isn't, the error is still legible.

An alternative would be to add the variant, which fails to parse in
clients that are not updated. (I.e. the error will just say "failed to
unmarshal error".) But the code would be cleaner, and the same as on
`main`.

## Test Plan

A test was added.

## Release Plan

- This should be released in a new SDK _first_, then the validators
should be updated.

## Links

- Related to https://github.com/linera-io/linera-protocol/issues/5023.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)