Autosigning on the Web (#5116)
## Motivation

Signing blocks on the Web can get tiresome when using wallets like
MetaMask or Dynamic, since we aren't integrated with their autosigning
features, and working with chains with hundreds of incoming messages
typically means accepting hundreds of signatures.

## Proposal

First connect to the ‘real’ wallet, and use it to sign a block that adds
a new local keypair as an additional owner of the chain. Then, replace
the original owner of the chain with this new temporary owner in the
wallet so that future blocks will be signed with that key instead. We
currently keep the wallet in memory on the Web so needn't worry about
overriding the real owner.

This is a slightly clumsy approach (and the application must be careful
to send any assets back to the original owner!) but should work as a
first step.

Implement a little supporting API to the Web client, then demo an
autosigning flow in the MetaMask counter demo.

## Test Plan

Tested by hand.

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK,
    - be released in a validator hotfix.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)