Autosigning improvements (#5126) (#5146)
## Motivation

@afck pointed out that having a lot of owners with non-zero weights
would slow the chain down in single-leader rounds.

@ma2bd pointed out that we need to make mutations using the original
owner.

## Proposal

Add a `weight` option to `addOwner` in an options object, and default it
to zero.

Add an optional `owner` option to `query` in an options object, and use
it to sign mutations. Also, document the above inline.

## Test Plan

Tested locally. Also, CI should test some of the API (but sadly not yet
as much as we'd like it to).

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK,

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)