[testnet] Follow-only chains (#5101) (#5114)
Backport of #5101, with backwards-compatible wallet.

## Motivation

All chains in the wallet are currently fully synchronized, including
their sender and publisher chains. But in use cases where the user only
wants to _read_ from that chain and not _write_ to it (i.e. never make a
block proposal, e.g. because they don't even own it), sender and
publisher chains are not needed, and neither are the chain manager
values.

## Proposal

Add a new listening mode and flag that marks chains as "follow-only".
For those chains, only their own confirmed blocks are synchronized.

The `follow-chain` and `forget-keys` commands automatically set that
flag, and `assign` unsets it.

In this backport, we also added code to read the old wallet format and
initialize the flag based on whether a chain has a keypair assigned.

## Test Plan

A test was added.

## Release Plan

- Should be released in a new SDK.

## Links

- PR to main: #5101
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)