Add liveness probes to proxy/shards (#5169) (#5182)
## Motivation

This is a backport of #5169.
Shards can occasionally become completely unresponsive due to, for
example, deadlocks or livelocks. When this happens, the pod can stay
"Running" but for example accepts no connections, requiring manual
intervention to detect and restart.

## Proposal

Add TCP liveness probes to proxy and shard pods so Kubernetes
automatically restarts unresponsive pods.

## Test Plan

Deployed a temporary network with these changes to make sure everything
seems to work